added toDto publisher

+++ src/main/java/kz/halykacademy/bookstore/repository/BookRepository.java
@@ -10,9 +10,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BookRepository extends JpaRepository<Books, Integer> {
+public interface BookRepository extends JpaRepository<Books, Long> {
 
     @Query(value = "SELECT  u FROM Books u WHERE u.title  LIKE %?1% ")
     List<Books> findByTitle(String title);
+
+
 
 }
